[EdgeTpu] Expand Cfg type for .edgetpucfg
- define OneCfg, EdgeTpuCfg type to ConfigSettings
- expand Cfg type
- make getter for ConfigSetting

ONE-vscode-DCO-1.0-Signed-off-by: Seongwon Im <imsw0529@gmail.com>
