+++ src/OneExplorer/ConfigSettings/OneConfigSetting.ts
@@ -19,14 +19,14 @@ import * as vscode from "vscode";
 import { Locator, LocatorRunner } from "../ArtifactLocator";
 import { ConfigSetting } from "../ConfigSetting";
 
-// export type OneCfg = {
-//   "one-import-tflite": CfgOneImportTflite;
-//   "one-import-onnx": CfgOneImportOnnx;
-//   "one-import-tf": CfgOneImportTf;
-// };
-// type CfgOneImportTflite = any;
-// type CfgOneImportOnnx = any;
-// type CfgOneImportTf = any;
+export type OneCfg = {
+  "one-import-tflite": CfgOneImportTflite;
+  "one-import-onnx": CfgOneImportOnnx;
+  "one-import-tf": CfgOneImportTf;
+};
+type CfgOneImportTflite = any;
+type CfgOneImportOnnx = any;
+type CfgOneImportTf = any;
 
 export class OneConfigSetting extends ConfigSetting {
   // TODO: make sections for updateBaseModelField method
